# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmSci E-Voting Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ApiError {
  error: string;
  /** @nullable */
  details?: string | null;
}

export type ElectionStatusPhase = typeof ElectionStatusPhase[keyof typeof ElectionStatusPhase];


export const ElectionStatusPhase = {
  setup: 'setup',
  registration: 'registration',
  voting: 'voting',
  results: 'results',
  audit: 'audit',
} as const;

export interface ElectionStatus {
  phase: ElectionStatusPhase;
  registrationOpen: boolean;
  votingOpen: boolean;
  /** @nullable */
  registrationStart?: string | null;
  /** @nullable */
  registrationEnd?: string | null;
  /** @nullable */
  votingStart?: string | null;
  /** @nullable */
  votingEnd?: string | null;
  electionName: string;
  totalStudents: number;
}

export interface ElectionSettings {
  id: number;
  electionName: string;
  /** @nullable */
  registrationStart?: string | null;
  /** @nullable */
  registrationEnd?: string | null;
  /** @nullable */
  votingStart?: string | null;
  /** @nullable */
  votingEnd?: string | null;
  showLiveResults: boolean;
  totalExpectedVoters: number;
}

export interface ElectionSettingsInput {
  electionName: string;
  /** @nullable */
  registrationStart?: string | null;
  /** @nullable */
  registrationEnd?: string | null;
  /** @nullable */
  votingStart?: string | null;
  /** @nullable */
  votingEnd?: string | null;
  showLiveResults: boolean;
  totalExpectedVoters: number;
}

export type VoterRegistrationInputLevel = typeof VoterRegistrationInputLevel[keyof typeof VoterRegistrationInputLevel];


export const VoterRegistrationInputLevel = {
  '100L': '100L',
  '200L': '200L',
  '300L': '300L',
  '400L': '400L',
  '500L': '500L',
  '600L': '600L',
} as const;

export interface VoterRegistrationInput {
  matricNumber: string;
  email: string;
  password: string;
  fullName: string;
  level: VoterRegistrationInputLevel;
}

export interface RegistrationResult {
  success: boolean;
  message: string;
  matricNumber: string;
  fullName: string;
  level: string;
}

export interface VoterLoginInput {
  matricNumber: string;
  password: string;
}

export interface VoterSession {
  id: number;
  matricNumber: string;
  fullName: string;
  level: string;
  hasVoted: boolean;
}

export interface Candidate {
  id: number;
  officeId: number;
  fullName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  level?: string | null;
}

export interface Office {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  displayOrder: number;
  candidates: Candidate[];
}

export interface Ballot {
  offices: Office[];
  voterName: string;
  voterLevel: string;
}

export interface BallotVote {
  officeId: number;
  candidateId: number;
}

export interface BallotSubmission {
  votes: BallotVote[];
}

export interface AdminLoginInput {
  email: string;
  password: string;
  /** @nullable */
  totpCode?: string | null;
}

export type AdminSessionRole = typeof AdminSessionRole[keyof typeof AdminSessionRole];


export const AdminSessionRole = {
  super_admin: 'super_admin',
  editor: 'editor',
  observer: 'observer',
} as const;

export interface AdminSession {
  id: number;
  email: string;
  name: string;
  role: AdminSessionRole;
}

export type AdminAccountRole = typeof AdminAccountRole[keyof typeof AdminAccountRole];


export const AdminAccountRole = {
  super_admin: 'super_admin',
  editor: 'editor',
  observer: 'observer',
} as const;

export interface AdminAccount {
  id: number;
  email: string;
  name: string;
  role: AdminAccountRole;
  createdAt: string;
  /** @nullable */
  lastLogin?: string | null;
}

export type AdminAccountInputRole = typeof AdminAccountInputRole[keyof typeof AdminAccountInputRole];


export const AdminAccountInputRole = {
  super_admin: 'super_admin',
  editor: 'editor',
  observer: 'observer',
} as const;

export interface AdminAccountInput {
  email: string;
  name: string;
  password: string;
  role: AdminAccountInputRole;
}

export interface LevelBreakdown {
  level: string;
  registered: number;
  voted: number;
}

export interface DashboardStats {
  totalRegistered: number;
  totalVoted: number;
  totalExpected: number;
  registrationRate: number;
  turnoutRate: number;
  byLevel: LevelBreakdown[];
  phase: string;
  registrationOpen: boolean;
  votingOpen: boolean;
}

export interface TurnoutTimelinePoint {
  hour: string;
  cumulative: number;
  delta: number;
}

export interface CandidateResult {
  candidateId: number;
  candidateName: string;
  voteCount: number;
  percentage: number;
  isWinner: boolean;
  isTie: boolean;
}

export interface OfficeResult {
  officeId: number;
  officeTitle: string;
  totalVotes: number;
  candidates: CandidateResult[];
}

export interface Voter {
  id: number;
  matricNumber: string;
  fullName: string;
  email: string;
  level: string;
  hasVoted: boolean;
  registrationTimestamp: string;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  voteTimestamp?: string | null;
}

export type VoterUpdateLevel = typeof VoterUpdateLevel[keyof typeof VoterUpdateLevel];


export const VoterUpdateLevel = {
  '100L': '100L',
  '200L': '200L',
  '300L': '300L',
  '400L': '400L',
  '500L': '500L',
  '600L': '600L',
} as const;

export interface VoterUpdate {
  fullName?: string;
  level?: VoterUpdateLevel;
}

export interface VoterListResponse {
  voters: Voter[];
  total: number;
  page: number;
  limit: number;
}

export interface OfficeInput {
  title: string;
  /** @nullable */
  description?: string | null;
  displayOrder: number;
}

export interface CandidateInput {
  officeId: number;
  fullName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  level?: string | null;
}

export type AuditLogActorType = typeof AuditLogActorType[keyof typeof AuditLogActorType];


export const AuditLogActorType = {
  student: 'student',
  admin: 'admin',
  system: 'system',
} as const;

export interface AuditLog {
  id: number;
  eventType: string;
  description: string;
  /** @nullable */
  actorId?: string | null;
  actorType: AuditLogActorType;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
  /** @nullable */
  metadata?: string | null;
}

export interface AuditLogResponse {
  logs: AuditLog[];
  total: number;
  page: number;
  limit: number;
}

export interface StudentRecord {
  id: number;
  matricNumber: string;
  email: string;
  fullName: string;
  level: string;
  isRegistered: boolean;
}

export interface StudentRecordListResponse {
  students: StudentRecord[];
  total: number;
  page: number;
  limit: number;
}

export type StudentRecordImportRecordsItem = {
  matricNumber: string;
  email: string;
  fullName: string;
  level: string;
};

export interface StudentRecordImport {
  records: StudentRecordImportRecordsItem[];
}

export interface ImportResult {
  imported: number;
  skipped: number;
  errors: number;
  message: string;
}

export type ListVotersParams = {
level?: string;
voted?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListAuditLogsParams = {
event?: string;
page?: number;
limit?: number;
};

export type ListStudentRecordsParams = {
search?: string;
level?: string;
page?: number;
limit?: number;
};