# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PharmSci E-Voting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/auth/login:
    post:
      operationId: loginAdmin
      tags: [admin-auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/auth/me:
    get:
      operationId: getAdminSession
      tags: [admin-auth]
      summary: Get current admin session
      responses:
        "200":
          description: Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Admin dashboard ─────────────────────────────────────────────────────────
...
  /admin/offices:
    get:
      operationId: listOffices
      tags: [admin-offices]
      summary: List all offices with their candidates
      responses:
        "200":
          description: Offices list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Office"
    post:
      operationId: createOffice
      tags: [admin-offices]
      summary: Create a new election office
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficeInput"
      responses:
        "201":
          description: Created office
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
  /admin/offices/{id}:
    put:
      operationId: updateOffice
      tags: [admin-offices]
      summary: Update an office
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficeInput"
      responses:
        "200":
          description: Updated office
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
    delete:
      operationId: deleteOffice
      tags: [admin-offices]
      summary: Delete an office
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/candidates:
    post:
      operationId: createCandidate
      tags: [admin-offices]
      summary: Add a candidate to an office
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /admin/candidates/{id}:
    put:
      operationId: updateCandidate
      tags: [admin-offices]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [admin-offices]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Admin results ────────────────────────────────────────────────────────────
...
  /admin/settings:
    get:
      operationId: getElectionSettings
      tags: [admin-settings]
      summary: Get election phase date/time settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionSettings"
    put:
      operationId: updateElectionSettings
      tags: [admin-settings]
      summary: Update election phase settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElectionSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionSettings"

  # ── Admin accounts ───────────────────────────────────────────────────────────
  /admin/accounts:
    get:
      operationId: listAdminAccounts
      tags: [admin-accounts]
      summary: List admin accounts (super admin only)
      responses:
        "200":
          description: Admin accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAccount"
    post:
      operationId: createAdminAccount
      tags: [admin-accounts]
      summary: Create admin account (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAccountInput"
      responses:
        "201":
          description: Created admin account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}:
    delete:
      operationId: deleteAdminAccount
      tags: [admin-accounts]
      summary: Delete admin account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Admin student records ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ApiError:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required: [error]

    # ── Election ────────────────────────────────────────────────────────────────
...
    ElectionSettings:
      type: object
      properties:
        id:
          type: integer
        electionName:
          type: string
        registrationStart:
          type: ["string", "null"]
        registrationEnd:
          type: ["string", "null"]
        votingStart:
          type: ["string", "null"]
        votingEnd:
          type: ["string", "null"]
        showLiveResults:
          type: boolean
        totalExpectedVoters:
          type: integer
      required: [id, electionName, showLiveResults, totalExpectedVoters]
    ElectionSettingsInput:
      type: object
      properties:
        electionName:
          type: string
        registrationStart:
          type: ["string", "null"]
        registrationEnd:
          type: ["string", "null"]
        votingStart:
          type: ["string", "null"]
        votingEnd:
          type: ["string", "null"]
        showLiveResults:
          type: boolean
        totalExpectedVoters:
          type: integer
      required: [electionName, showLiveResults, totalExpectedVoters]

    # ── Student auth ────────────────────────────────────────────────────────────
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
        officeId:
          type: integer
        fullName:
          type: string
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
      required: [id, officeId, fullName]
    Office:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        displayOrder:
          type: integer
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
      required: [id, title, displayOrder, candidates]
...
    AdminLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        totpCode:
          type: ["string", "null"]
      required: [email, password]
    AdminSession:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, editor, observer]
      required: [id, email, name, role]
    AdminAccount:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, editor, observer]
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
      required: [id, email, name, role, createdAt]
    AdminAccountInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [super_admin, editor, observer]
      required: [email, name, password, role]

    # ── Dashboard ───────────────────────────────────────────────────────────────
...
    OfficeInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required: [title, displayOrder]
    CandidateInput:
      type: object
      properties:
        officeId:
          type: integer
        fullName:
          type: string
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
      required: [officeId, fullName]

    # ── Audit ────────────────────────────────────────────────────────────────────
...