# lib/api-zod/src/generated/types/adminAccount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmSci E-Voting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAccountRole } from './adminAccountRole';

export interface AdminAccount {
  id: number;
  email: string;
  name: string;
  role: AdminAccountRole;
  createdAt: string;
  /** @nullable */
  lastLogin?: string | null;
}